Fix some feature envy
