+++ src/main/java/no/elg/ii/inventory/InventoryService.java
@@ -42,14 +42,8 @@ import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
 import net.runelite.api.InventoryID;
 import net.runelite.api.ItemContainer;
-import net.runelite.api.events.BeforeRender;
 import net.runelite.api.widgets.Widget;
-import net.runelite.client.eventbus.Subscribe;
-import no.elg.ii.inventory.slot.InventorySlot;
-import no.elg.ii.service.EnsureWidgetStateService;
-import no.elg.ii.service.WidgetService;
 import no.elg.ii.util.IndexedWidget;
-import no.elg.ii.util.WidgetUtils;
 
 @Singleton
 @Slf4j
@@ -57,10 +51,6 @@ public class InventoryService {
 
   @Inject
   private Client client;
-  @Inject
-  private EnsureWidgetStateService ensureWidgetStateService;
-  @Inject
-  private WidgetService widgetService;
 
   @Nullable
   public ItemContainer getCurrentInventoryContainer() {
@@ -89,22 +79,5 @@ public class InventoryService {
   public final Stream<IndexedWidget> getAllOpenInventoryWidgets() {
     return getOpenWidgetItemContainer()
       .flatMap(container -> Streams.mapWithIndex(Arrays.stream(container.getDynamicChildren()), indexWidget));
-  }
-
-  public boolean isDifferent(Widget widget, InventorySlot slot) {
-    return slot.hasValidItemId() && !WidgetUtils.isEmpty(widget)
-      && (widget.getItemId() != slot.getItemId()
-      || widget.getItemQuantity() != slot.getQuantity()
-      || widget.getOpacity() != slot.getOpacity());
-  }
-
-  public void setWidgetFromSlot(Widget widget, InventorySlot slot) {
-    widgetService.updateVisibleWidget(widget, slot.getItemId(), slot.getQuantity());
-    widgetService.setOpacity(widget, slot.getOpacity(), true);
-  }
-
-  @Subscribe
-  public void onBeforeRender(BeforeRender event) {
-    ensureWidgetStateService.forceWidgetState(this::isDifferent, this::setWidgetFromSlot);
   }
 }
